av1_grain

Module create

Structs

NoiseGenArgs

Settings and video data defining how to generate the film grain params.

Enums

TransferFunction

Constants

BT1886_BLACKPOINT

BT1886_GAMMA

BT1886_WHITEPOINT

PQ_C1

PQ_C2

PQ_C3

PQ_M1

PQ_M2

Functions

bt1886_alpha

The variable for user gain: 伪 = (Lw^(1/位) - Lb^(1/位)) ^ 位

bt1886_beta

The variable for user black level lift: 尾 = Lb^(1/位) / (Lw^(1/位) - Lb^(1/位))

bt1886_inv_blackpoint

bt1886_inv_whitepoint

generate_luma_noise_points

generate_photon_noise_params

Generates a set of photon noise parameters for a segment of video given a set of args.

write_film_grain_segment

write_grain_table

Write a set of generated film grain params to a table file, using the standard film grain table format supported by aomenc, rav1e, and svt-av1.