av1_grain

Function write_grain_table

pub fn write_grain_table<P: AsRef<Path>>(
    filename: P,
    params: &[GrainTableSegment],
) -> Result<()>

Write a set of generated film grain params to a table file, using the standard film grain table format supported by aomenc, rav1e, and svt-av1.

Errors

If the output file cannot be written to