Crate av1_grain

Re-exports

pub use v_frame;

Structs

DiffGenerator

GrainTableSegment

Specifies parameters for enabling decoder-side grain synthesis for a segment of video from start_time to end_time.

NoiseGenArgs

Settings and video data defining how to generate the film grain params.

Enums

TransferFunction

Constants

DEFAULT_GRAIN_SEED

A randomly generated u16 to be used as a starting random seed for grain synthesis. The idea behind using a constant random seed is so that encodes are deterministic and reproducible.

NUM_UV_COEFFS

The max number of coefficients per chroma plane for grain synthesis

NUM_UV_POINTS

The max number of scaling points per chroma plane for grain synthesis

NUM_Y_COEFFS

The max number of luma coefficients for grain synthesis

NUM_Y_POINTS

The max number of luma scaling points for grain synthesis

Functions

generate_photon_noise_params

Generates a set of photon noise parameters for a segment of video given a set of args.

parse_grain_table

This file has the implementation details of the grain table.

write_grain_table

Write a set of generated film grain params to a table file, using the standard film grain table format supported by aomenc, rav1e, and svt-av1.

Type Aliases

ScalingPoints